# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "",
        "contact": {
            "email": "afriantomahmud@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/unit": {
            "get": {
                "tags": [
                    "Units"
                ],
                "summary": "Get a list of units",
                "description": "Returns a list of units",
                "operationId": "7335d87c4a3157fe50761cf91d300722",
                "responses": {
                    "200": {
                        "description": "Successful response: List of units",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Data Units"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UnitResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Units"
                ],
                "summary": "Store a new unit",
                "description": "Store a new unit in database",
                "operationId": "302327015261cb20d1a667f4b5b4c63e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UnitRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response: List of units",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Data Units"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UnitResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/unit/{id}": {
            "get": {
                "tags": [
                    "Units"
                ],
                "summary": "Get a single of units",
                "description": "Returns a single of units by ID",
                "operationId": "5c3d4f307f1777efebfcf2772f04cb94",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the unit to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response: List of units",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Data Units"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UnitResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Units"
                ],
                "summary": "Update a new unit",
                "description": "Update a new unit in database",
                "operationId": "15fe2aa97c5f4205f164652072d9e9cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the unit to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UnitRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response: List of units",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Data Units"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UnitResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Units"
                ],
                "summary": "Delete a new unit",
                "description": "Delete a new unit in database",
                "operationId": "550f0ff4b8d51a1c33bd3a3d08e1efb1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the unit to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Data Units"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UnitRequest": {
                "title": "Unit Request",
                "description": "Request structure for unit",
                "required": [
                    "code",
                    "name"
                ],
                "properties": {
                    "code": {
                        "description": "code",
                        "type": "string"
                    },
                    "name": {
                        "description": "name",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UnitResource": {
                "title": "Unit Resource",
                "description": "Unit Resource",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the unit",
                        "type": "string"
                    },
                    "code": {
                        "description": "The code of the post",
                        "type": "string"
                    },
                    "name": {
                        "description": "The name of the post",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}